+++ spec/anchor.spec.js
@@ -1,38 +1,31 @@
 /*global MediumEditor, describe, it, expect, spyOn,
          afterEach, beforeEach, selectElementContents,
-         jasmine, fireEvent, console*/
+         jasmine, fireEvent, console, tearDown*/
 
 describe('Anchor Button TestCase', function () {
     'use strict';
 
     beforeEach(function () {
-        jasmine.Clock.useMock();
-        this.body = document.getElementsByTagName('body')[0];
+        jasmine.clock().install();
         this.el = document.createElement('div');
         this.el.className = 'editor';
         this.el.innerHTML = 'lorem ipsum';
-        this.body.appendChild(this.el);
+        document.body.appendChild(this.el);
     });
 
     afterEach(function () {
-        var elements = document.querySelectorAll('.medium-editor-toolbar'),
-            i,
-            sel = window.getSelection();
-        for (i = 0; i < elements.length; i += 1) {
-            this.body.removeChild(elements[i]);
-        }
-        this.body.removeChild(this.el);
-        sel.removeAllRanges();
+        tearDown(this.el);
+        jasmine.clock().uninstall();
     });
 
     describe('Click', function () {
         it('should display the anchor form when toolbar is visible', function () {
-            spyOn(MediumEditor.prototype, 'showAnchorForm').andCallThrough();
+            spyOn(MediumEditor.prototype, 'showAnchorForm').and.callThrough();
             var button,
                 editor = new MediumEditor('.editor');
             selectElementContents(editor.elements[0]);
             fireEvent(editor.elements[0], 'mouseup');
-            jasmine.Clock.tick(1);
+            jasmine.clock().tick(1);
             button = editor.toolbar.querySelector('[data-element="a"]');
             fireEvent(button, 'click');
             expect(editor.toolbarActions.style.display).toBe('none');
@@ -41,12 +34,12 @@ describe('Anchor Button TestCase', function () {
         });
 
         it('should display the toolbar actions when anchor form is visible', function () {
-            spyOn(MediumEditor.prototype, 'showToolbarActions').andCallThrough();
+            spyOn(MediumEditor.prototype, 'showToolbarActions').and.callThrough();
             var button,
                 editor = new MediumEditor('.editor');
             selectElementContents(editor.elements[0]);
             fireEvent(editor.elements[0], 'mouseup');
-            jasmine.Clock.tick(1);
+            jasmine.clock().tick(1);
             button = editor.toolbar.querySelector('[data-element="a"]');
             editor.anchorForm.style.display = 'block';
             fireEvent(button, 'click');
@@ -56,13 +49,13 @@ describe('Anchor Button TestCase', function () {
         });
 
         it('should unlink when selection is a link', function () {
-            spyOn(document, 'execCommand').andCallThrough();
+            spyOn(document, 'execCommand').and.callThrough();
             this.el.innerHTML = '<a href="#">link</a>';
             var button,
                 editor = new MediumEditor('.editor');
             selectElementContents(editor.elements[0]);
             fireEvent(editor.elements[0], 'mouseup');
-            jasmine.Clock.tick(1);
+            jasmine.clock().tick(1);
             button = editor.toolbar.querySelector('[data-element="a"]');
             fireEvent(button, 'click');
             expect(this.el.innerHTML, 'link');
@@ -73,7 +66,7 @@ describe('Anchor Button TestCase', function () {
 
     describe('Link Creation', function () {
         it('should create a link when user presses enter', function () {
-            spyOn(MediumEditor.prototype, 'createLink').andCallThrough();
+            spyOn(MediumEditor.prototype, 'createLink').and.callThrough();
             var editor = new MediumEditor('.editor'),
                 button,
                 input;
@@ -88,7 +81,9 @@ describe('Anchor Button TestCase', function () {
         });
 
         it('should add target="_blank" when respective option is set to true', function () {
-            var editor = new MediumEditor('.editor', {targetBlank: true}),
+            var editor = new MediumEditor('.editor', {
+                targetBlank: true
+            }),
                 input = editor.anchorForm.querySelector('input');
             selectElementContents(editor.elements[0]);
             input.value = 'http://test.com';
@@ -99,7 +94,7 @@ describe('Anchor Button TestCase', function () {
 
     describe('Cancel', function () {
         it('should close the link form when user clicks on cancel', function () {
-            spyOn(MediumEditor.prototype, 'showToolbarActions').andCallThrough();
+            spyOn(MediumEditor.prototype, 'showToolbarActions').and.callThrough();
             var editor = new MediumEditor('.editor'),
                 button,
                 cancel;