+++ spec/elements.spec.js
@@ -1,24 +1,19 @@
 /*global MediumEditor, describe, it, expect, spyOn,
-         afterEach, beforeEach*/
+         afterEach, beforeEach, tearDown*/
 
 describe('Elements TestCase', function () {
     'use strict';
 
     describe('Initialization', function () {
         beforeEach(function () {
-            this.body = document.getElementsByTagName('body')[0];
+            document.body = document.getElementsByTagName('body')[0];
             this.el = document.createElement('div');
             this.el.className = 'editor';
-            this.body.appendChild(this.el);
+            document.body.appendChild(this.el);
         });
 
         afterEach(function () {
-            var elements = document.querySelectorAll('.medium-editor-toolbar'),
-                i;
-            for (i = 0; i < elements.length; i += 1) {
-                this.body.removeChild(elements[i]);
-            }
-            this.body.removeChild(this.el);
+            tearDown(this.el);
         });
 
         it('should set element contenteditable attribute to true', function () {