+++ spec/selection.spec.js
@@ -1,50 +1,43 @@
 /*global MediumEditor, describe, it, expect, spyOn,
          afterEach, beforeEach, fireEvent, waits,
-         jasmine, selectElementContents*/
+         jasmine, selectElementContents, tearDown*/
 
 describe('Selection TestCase', function () {
     'use strict';
 
     beforeEach(function () {
-        this.body = document.getElementsByTagName('body')[0];
         this.el = document.createElement('div');
         this.el.className = 'editor';
         this.el.innerHTML = 'lorem ipsum';
-        this.body.appendChild(this.el);
+        document.body.appendChild(this.el);
     });
 
     afterEach(function () {
-        var elements = document.querySelectorAll('.medium-editor-toolbar'),
-            i,
-            sel = window.getSelection();
-        for (i = 0; i < elements.length; i += 1) {
-            this.body.removeChild(elements[i]);
-        }
-        this.body.removeChild(this.el);
-        sel.removeAllRanges();
+        tearDown(this.el);
+        jasmine.clock().uninstall();
     });
 
     describe('CheckSelection', function () {
         it('should check for selection on mouseup event', function () {
-            jasmine.Clock.useMock();
+            jasmine.clock().install();
             spyOn(MediumEditor.prototype, 'checkSelection');
             var editor = new MediumEditor('.editor');
             fireEvent(editor.elements[0], 'mouseup');
-            jasmine.Clock.tick(1);
+            jasmine.clock().tick(1);
             expect(editor.checkSelection).toHaveBeenCalled();
         });
 
         it('should check for selection on keyup', function () {
-            jasmine.Clock.useMock();
+            jasmine.clock().install();
             spyOn(MediumEditor.prototype, 'checkSelection');
             var editor = new MediumEditor('.editor');
             fireEvent(editor.elements[0], 'keyup');
-            jasmine.Clock.tick(1);
+            jasmine.clock().tick(1);
             expect(editor.checkSelection).toHaveBeenCalled();
         });
 
         it('should do nothing when keepToolbarAlive is true', function () {
-            spyOn(window, 'getSelection').andCallThrough();
+            spyOn(window, 'getSelection').and.callThrough();
             var editor = new MediumEditor('.editor');
             editor.keepToolbarAlive = true;
             editor.checkSelection();
@@ -53,9 +46,9 @@ describe('Selection TestCase', function () {
 
         describe('When keepToolbarAlive is false', function () {
             it('should hide the toolbar if selection is empty', function () {
-                spyOn(MediumEditor.prototype, 'setToolbarPosition').andCallThrough();
-                spyOn(MediumEditor.prototype, 'setToolbarButtonStates').andCallThrough();
-                spyOn(MediumEditor.prototype, 'showToolbarActions').andCallThrough();
+                spyOn(MediumEditor.prototype, 'setToolbarPosition').and.callThrough();
+                spyOn(MediumEditor.prototype, 'setToolbarButtonStates').and.callThrough();
+                spyOn(MediumEditor.prototype, 'showToolbarActions').and.callThrough();
                 var editor = new MediumEditor('.editor');
                 editor.toolbar.style.display = 'block';
                 editor.toolbar.classList.add('medium-editor-toolbar-active');
@@ -69,18 +62,18 @@ describe('Selection TestCase', function () {
 
             it('should show the toolbar when something is selected', function () {
                 var editor = new MediumEditor('.editor');
-                jasmine.Clock.useMock();
+                jasmine.clock().install();
                 expect(editor.toolbar.classList.contains('medium-editor-toolbar-active')).toBe(false);
                 selectElementContents(this.el);
                 editor.checkSelection();
-                jasmine.Clock.tick(501);
+                jasmine.clock().tick(501);
                 expect(editor.toolbar.classList.contains('medium-editor-toolbar-active')).toBe(true);
             });
 
             it('should update toolbar position and button states when something is selected', function () {
-                spyOn(MediumEditor.prototype, 'setToolbarPosition').andCallThrough();
-                spyOn(MediumEditor.prototype, 'setToolbarButtonStates').andCallThrough();
-                spyOn(MediumEditor.prototype, 'showToolbarActions').andCallThrough();
+                spyOn(MediumEditor.prototype, 'setToolbarPosition').and.callThrough();
+                spyOn(MediumEditor.prototype, 'setToolbarButtonStates').and.callThrough();
+                spyOn(MediumEditor.prototype, 'showToolbarActions').and.callThrough();
                 var editor = new MediumEditor('.editor');
                 selectElementContents(this.el);
                 editor.checkSelection();